WIP: Preliminary ArFS-drive indexing code operational.
The way the HTML-output is currently added is a quick hack - I just
wanted to quickly get my hands onto a HTML drive listing with links.

The output-processing code needs to be moved into its own file(s),
perhaps making a class that handles it.

But for now, I think I'll just improve the resulting page and
set up a small docker service with automatic updates to get
people exploring the public drives.