Things have been done.
Can now retrieve raw data of transactions (ie. path manifests).
Listing also now works properly, replacing entries if their
block timestamp is greater than the one in the stored entry.

The fetch is in need for serious optimization, though.
Currently it fetches newest metadata-tx for each file it updates,
causing a lot of overhead with those queries for one transaction.

Should just fetch ArFS-entries for the entire address, or possibly
limit to the chosen drive + add a flags to fetch one directory
at a time, or one file at a time. These shouldn't be needed, but
the reason I want to keep them that way is that the weave can
really bloat in time, addresses and drives containing a huge amount of
transactions, so I feel it's good to retain the option to do smaller
fetches at the time.

Another thing I'm going to try out is doing the fetches asynchronously,
maybe processing something like 10 entries per time when going through
a directory.

Some experimentation required.
But for now, it's time to get some sleep.

Silanael signing out.