FIX: Recursive directory updating now prevents eternal looping.
It keeps track of visited Folder-Ids, ignoring ones present
in the list. This is to avoid a hypothetical attack of someone
making a drive with a folder that pointed to the folder that
contained it, or to the root folder or so.

This is untested still. Arweave is currently heavily congested,
its mempool having 10711 transactions pending so a bit difficult
to do real-world testing.. A task for tomorrow.