CHANGE: Default concurrent delay upped to 200ms.
The former value of 50ms worked well for my connection but turned
out to be too fast for some. 200ms should be a safe default - one
can still manually drop it to 50 with --concurrent-ms 50 before
I make it dynamic or try a different query amount -based approach
in future versions.