Version 0.8.5 Alpha
"Dying Light"

This is a pre-release version that is not intended for production.

That said, many of its functions are operational - VERIFY
can be used to pull drive listings (filename-DTXID CSVs), list
successful and failed file uploads (along with missing files with
the NUMERIC mode). The INFO-command gives detailed information about
transactions and ArFS-entities, while STATUS does some further analysis
on the state of drives, files and folders, finding orphaned (ie. folderless)
ArFS-entities. Furthermore, 'LIST DRIVES <drive-id> DEEP' should find broken
drives resulted in the drive metadata failing to be mined.

This is still a work in progress.