CHANGE: Concurrent code altered.
Ditched the multi-fetch functionality and the Manager altogether,
making a system that works well with Promise.all.

Furthermore, ArFSEntity.FetchContentEntities now also fetches
the newest meta-TX for the contained entities to ensure that
they haven't been moved out of the folder, if a folder is
in question. Drive query does no such checks.