CHANGE: Default fields are now set in SARTObject.
The config-keys now act the same way as manually given fields would,
groups and +- -modifiers usable with them. As such, the group
'DEFAULT' can't be used to cause a recursive loop (not that it could
lead into that even with the previous implementation).

Furthermore, added support for a '+' field prefix, allowing the user
to give a field list such as "+Fee_Winston,+BlockID" resulting in
the fields being appended to defaults instead of used as a field list.

I feel that the field select system is in a rather versatile condition.
Next thing to do will be to re-add entry mode output, implement proper
listing/display of recursive items and get the other OutputFormats up,
perhaps then I can finally start on re-implementing ArFS-listing.

The exhaustion has been strong, causing each line of code written
to be a nearly overwhelming challenge. Yet I'm determined to push
through it, proving that I have what it takes to achieve something
even despite of this crippling feverishness and exhaustion, before
ruling out the last possible medical culprit in a surgery I have
scheduled in March. The odds are that it will not improve
the situation, but in case it will, I will take this opportunity
to measure and immortalize my true power by getting SART into
a fully operational and useful condition. Yet, this tool is a mere
stepping stone into projects of far greater magnitude - I will
most certainly be producing some FHC/FS -content before the
deadline I've set for myself. Perhaps even start prototyping
FHCU/FSU.. The first game was supposed to be a mobile arcade
game (that I've had a working playable prototype for like 8
years now), yet I've always envisioned it to culminate into
an universe that will live its own live, continuing to exist
long after I'm gone - Arweave is absolutely perfect for this.